Finalização do CRUD de Empresa e Pessoa

+++ src/main/java/com/javaee/prince/trabalhofinaljee/bootstrap/ApplicationBootstrap.java
@@ -40,7 +40,7 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
 			loadAcoes();
 		}
 		
-/*		if (empresaRepository.count() == 0L) 
+		if (empresaRepository.count() == 0L) 
 		{
 			empresaRepository.deleteAll();
 			loadEmpresas();
@@ -50,7 +50,7 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
 		{
 			pessoaRepository.deleteAll();
 			loadPessoas();
-		}*/
+		}
 	}
 	
 	private void loadAcoes() 
@@ -59,7 +59,7 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
 		String dataInicial = java.text.DateFormat.getDateInstance(DateFormat.MEDIUM).format(dataAtual);
 		
         Acao acaoCemig = new Acao();
-        acaoCemig.setAcaoNome("CMIG4");
+        acaoCemig.setName("CMIG4");
         acaoCemig.setValorInicial(14.43f);
         acaoCemig.setValorAtual(0);
         acaoCemig.setIdEmpresa(1);
@@ -68,7 +68,7 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
         acaoRepository.save(acaoCemig);
         
         Acao acaoMagazineLuiza = new Acao();
-        acaoMagazineLuiza.setAcaoNome("MGLU3");
+        acaoMagazineLuiza.setName("MGLU3");
         acaoMagazineLuiza.setValorInicial(181.63f);
         acaoMagazineLuiza.setValorAtual(0);
         acaoMagazineLuiza.setIdEmpresa(2);
@@ -77,7 +77,7 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
         acaoRepository.save(acaoMagazineLuiza);
 
         Acao acaoOracle = new Acao();
-        acaoOracle.setAcaoNome("ORCL34");
+        acaoOracle.setName("ORCL34");
         acaoOracle.setValorInicial(204.73f);
         acaoOracle.setValorAtual(0);
         acaoOracle.setIdEmpresa(3);
@@ -86,42 +86,42 @@ public class ApplicationBootstrap implements ApplicationListener<ContextRefreshe
         acaoRepository.save(acaoOracle);
 	}
 	
-/*	private void loadEmpresas() 
+	private void loadEmpresas() 
 	{
         Empresa empresa01 = new Empresa();
-        empresa01.setEmpresaNome("Cemig");
+        empresa01.setName("Cemig");
         empresa01.setEmpresaId(1);
         empresaRepository.save(empresa01);
         
         Empresa empresa02 = new Empresa();
-        empresa02.setEmpresaNome("Magazine Luiza");
+        empresa02.setName("Magazine Luiza");
         empresa01.setEmpresaId(2);
         empresaRepository.save(empresa02);
         
         Empresa empresa03 = new Empresa();
-        empresa03.setEmpresaNome("Oracle");
+        empresa03.setName("Oracle");
         empresa03.setEmpresaId(3);
         empresaRepository.save(empresa03);
-	}*/
+	}
 	
-/*	private void loadPessoas() 
+	private void loadPessoas() 
 	{
         Pessoa pessoa01 = new Pessoa();
-        pessoa01.setPessoaNome("João");
+        pessoa01.setName("João");
         pessoa01.setPessoaEmail("joao.silva@gmail.com");
         pessoa01.setPessoaCpf("83128339007");
         pessoaRepository.save(pessoa01);
         
         Pessoa pessoa02 = new Pessoa();
-        pessoa02.setPessoaNome("Luiz");
+        pessoa02.setName("Luiz");
         pessoa02.setPessoaEmail("luiz.josue@gmail.com");
         pessoa02.setPessoaCpf("53182603060");
         pessoaRepository.save(pessoa02);
         
         Pessoa pessoa03 = new Pessoa();
-        pessoa03.setPessoaNome("Samuel");
+        pessoa03.setName("Samuel");
         pessoa03.setPessoaEmail("samuel.junior@gmail.com");
         pessoa03.setPessoaCpf("24502762032");
         pessoaRepository.save(pessoa03);
-	}*/
+	}
 }
